requirements.txt
python-dotenv
Ponytile
langchain
langchain_openai 
requests
